log_management/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
elasticsearch-dsl==7.4.1
web3==6.11.1
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1